Add sphinx-1.5
Strips out the conversion for moving from <1.1.3-r4, as these versions are
old enough to not be in the git history (2015-08-08).

Removes 'grammar generation' by loading sphinx.pycode.__init__, as
filenames in that file are out of date (sphinx/environment.py ->
sphinx/environment/__init__.py), and not mentioned in the official install
docs anywhere.